// src/api/generated/schemas/disciplineResponseDtoListApiResponse.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * NBPEM API 1.0
 * OpenAPI spec version: v1
 */
import type { DisciplineResponseDto } from './disciplineResponseDto';
import type { DisciplineResponseDtoListApiResponseErrors } from './disciplineResponseDtoListApiResponseErrors';

export interface DisciplineResponseDtoListApiResponse {
  /** @nullable */
  message?: string | null;
  status?: number;
  /** @nullable */
  data?: DisciplineResponseDto[] | null;
  /** @nullable */
  errors?: DisciplineResponseDtoListApiResponseErrors;
}