// src/api/generated/schemas/documentGroupResponseDtoListApiResponse.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * NBPEM API 1.0
 * OpenAPI spec version: v1
 */
import type { DocumentGroupResponseDto } from './documentGroupResponseDto';
import type { DocumentGroupResponseDtoListApiResponseErrors } from './documentGroupResponseDtoListApiResponseErrors';

export interface DocumentGroupResponseDtoListApiResponse {
  /** @nullable */
  message?: string | null;
  status?: number;
  /** @nullable */
  data?: DocumentGroupResponseDto[] | null;
  /** @nullable */
  errors?: DocumentGroupResponseDtoListApiResponseErrors;
}