// src/api/generated/schemas/documentTypeResponseDtoListApiResponse.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * NBPEM API 1.0
 * OpenAPI spec version: v1
 */
import type { DocumentTypeResponseDto } from './documentTypeResponseDto';
import type { DocumentTypeResponseDtoListApiResponseErrors } from './documentTypeResponseDtoListApiResponseErrors';

export interface DocumentTypeResponseDtoListApiResponse {
  /** @nullable */
  message?: string | null;
  status?: number;
  /** @nullable */
  data?: DocumentTypeResponseDto[] | null;
  /** @nullable */
  errors?: DocumentTypeResponseDtoListApiResponseErrors;
}