// src/api/generated/schemas/facilityCodeResponseDtoListApiResponse.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * NBPEM API 1.0
 * OpenAPI spec version: v1
 */
import type { FacilityCodeResponseDto } from './facilityCodeResponseDto';
import type { FacilityCodeResponseDtoListApiResponseErrors } from './facilityCodeResponseDtoListApiResponseErrors';

export interface FacilityCodeResponseDtoListApiResponse {
  /** @nullable */
  message?: string | null;
  status?: number;
  /** @nullable */
  data?: FacilityCodeResponseDto[] | null;
  /** @nullable */
  errors?: FacilityCodeResponseDtoListApiResponseErrors;
}